requirements.txt
# LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-core>=0.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
fastmcp>=0.2.0
ultralytics>=8.0.0
transformers>=4.35.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
pydantic>=2.0.0
tqdm
pandas
chromadb
sentence_transformers
streamlit